- Implemento retrofit y traigo las categorias para mostrar!

+++ app/src/main/java/com/easypick/easypick/Locales.kt
@@ -1,3 +1,3 @@
 package com.easypick.easypick
 
-class Locales(var titulo:String, var detalle:String, var foto: Int, var id: Int)
+class Locales(var titulo:String, var detalle:String, var foto: Int, var id: Long)
